All DailyPrice to Integer

+++ grupo2-backend/src/main/java/com/integrator/group2backend/dto/ProductViewDTO.java
@@ -19,7 +19,7 @@ public class ProductViewDTO {
     private Integer beds;
     private Integer bathrooms;
     private Integer guests;
-    private Float dailyPrice;
+    private Integer dailyPrice;
     private String address;
     private Integer number;
     private Integer floor;
